% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap.r
\name{create_heatmap}
\alias{create_heatmap}
\title{Title}
\usage{
create_heatmap(data = NULL, type = "plotly")
}
\arguments{
\item{data}{Data to be plotted}

\item{type}{R-package to generate the plot. c("plotly", "heatmaply")}
}
\value{
A plotly heatmap
}
\description{
Title
}
