% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.r
\name{dotplot}
\alias{dotplot}
\title{Make a plot}
\usage{
dotplot(data_to_plot = NULL, all_data = NULL, ref_line = 30)
}
\arguments{
\item{data_to_plot}{Data to plot}

\item{all_data}{All the data (used to make median line)}

\item{ref_line}{Reference line}
}
\value{
A plot
}
\description{
Make a plot
}
